refactor: change OfflineArea conversion functions to extensions
